Vector Simplification + Added Quad2D + Vector2D

+++ src/main/java/sonar/logistics/client/gsi/properties/BorderProperty.java
@@ -5,15 +5,15 @@ import net.minecraft.util.math.Vec3d;
 /**convenience class, a simple float value which can be sized by percentage or be an absolute size*/
 public class BorderProperty {
 
-    public float value;
+    public double value;
     public boolean percentage;
 
-    public BorderProperty(float value, boolean percentage){
+    public BorderProperty(double value, boolean percentage){
         this.value = value;
         this.percentage = percentage;
     }
 
-    public float getRenderSize(float maxSize){
+    public double getRenderSize(double maxSize){
         if(!percentage){
             return value;
         }
